+++ core/src/main/java/nl/nn/adapterframework/compression/ZipWriter.java
@@ -1,5 +1,5 @@
 /*
-   Copyright 2013 Nationale-Nederlanden, 2020 WeAreFrank!
+   Copyright 2013 Nationale-Nederlanden, 2020, 2022 WeAreFrank!
 
    Licensed under the Apache License, Version 2.0 (the "License");
    you may not use this file except in compliance with the License.
@@ -15,6 +15,8 @@
 */
 package nl.nn.adapterframework.compression;
 
+import java.io.FileNotFoundException;
+import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
@@ -22,28 +24,47 @@ import java.util.zip.CRC32;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipOutputStream;
 
+import javax.servlet.http.HttpServletResponse;
+
 import org.apache.commons.lang3.StringUtils;
 import org.apache.logging.log4j.Logger;
 
+import lombok.Getter;
+import lombok.Setter;
+import nl.nn.adapterframework.collection.CollectionActor.Action;
+import nl.nn.adapterframework.collection.CollectionException;
+import nl.nn.adapterframework.collection.ICollector;
+import nl.nn.adapterframework.configuration.ConfigurationException;
+import nl.nn.adapterframework.core.IForwardTarget;
 import nl.nn.adapterframework.core.PipeLineSession;
+import nl.nn.adapterframework.core.TimeoutException;
+import nl.nn.adapterframework.parameters.Parameter;
+import nl.nn.adapterframework.parameters.ParameterList;
+import nl.nn.adapterframework.parameters.ParameterValue;
+import nl.nn.adapterframework.parameters.ParameterValueList;
 import nl.nn.adapterframework.stream.Message;
+import nl.nn.adapterframework.stream.MessageOutputStream;
 import nl.nn.adapterframework.util.LogUtil;
 import nl.nn.adapterframework.util.Misc;
+import nl.nn.adapterframework.util.StreamUtil;
 
 /**
  * Helper class to create Zip archives.
- * 
+ *
  * @author  Gerrit van Brakel
  * @since   4.9.10
  */
-public class ZipWriter {
-	protected Logger log = LogUtil.getLogger(this);
+public class ZipWriter implements ICollector<IZipWritingElement> {
+	protected static Logger log = LogUtil.getLogger(ZipWriter.class);
 
-	private ZipOutputStream zipoutput;
+	public static final String PARAMETER_FILENAME="filename";
+	public static final String PARAMETER_CONTENTS="contents";
+
+	private @Getter ZipOutputStream zipoutput;
 	private boolean entryOpen=false;
-	private boolean closeOnExit;
+	private @Getter @Setter boolean closeOnExit;
 
-	private ZipWriter(OutputStream resultStream, boolean closeOnExit) {
+	public ZipWriter(OutputStream resultStream, boolean closeOnExit) {
 		super();
 		this.closeOnExit=closeOnExit;
 		zipoutput=new ZipOutputStream(resultStream);
@@ -56,8 +77,70 @@ public class ZipWriter {
 	public static ZipWriter createZipWriter(PipeLineSession session, String handlekey, OutputStream resultStream, boolean closeOnExit) {
 		ZipWriter handle=new ZipWriter(resultStream,closeOnExit);
 		session.put(handlekey,handle);
-		if (handle.log.isDebugEnabled()) handle.log.debug(handle.getLogPrefix(handlekey)+"opened new zipstream");
+		handle.log.debug("opened new zipstream [{}]", handlekey);
 		return handle;
+	}
+
+	public static void configure(Action action, ParameterList parameterList) throws ConfigurationException {
+		switch (action) {
+			case OPEN:
+				break;
+			case WRITE:
+			case LAST:
+			case STREAM:
+				Parameter filenameParameter=parameterList.findParameter(PARAMETER_FILENAME);
+				Parameter contentsParameter=parameterList.findParameter(PARAMETER_CONTENTS);
+				if (filenameParameter==null && contentsParameter==null) {
+					throw new ConfigurationException("parameter '"+PARAMETER_FILENAME+"' or parameter '"+PARAMETER_CONTENTS+"' is required");
+				}
+				break;
+			case CLOSE:
+				if (parameterList.findParameter(PARAMETER_FILENAME)!=null) {
+					throw new ConfigurationException("with action ["+action+"] parameter '"+PARAMETER_FILENAME+"' cannot not be configured");
+				}
+				break;
+			default:
+				throw new ConfigurationException("unknwon action ["+action+"]");
+		}
+	}
+
+	public static ZipWriter openCollection(Message message, PipeLineSession session, ParameterValueList pvl, IZipWritingElement writingElement) throws CollectionException {
+		log.debug("opening new zipstream");
+		OutputStream resultStream=null;
+		Object input=message.asObject();
+		if (input==null) {
+			throw new CollectionException("input cannot be null, must be OutputStream, HttpResponse or String containing filename");
+		}
+		if (input instanceof OutputStream) {
+			resultStream=(OutputStream)input;
+		} else if (input instanceof HttpServletResponse) {
+			ParameterValue pv=pvl.get(PARAMETER_FILENAME);
+			if (pv==null) {
+				throw new CollectionException("parameter '"+PARAMETER_FILENAME+"' not found, but required if stream is HttpServletResponse");
+			}
+			String filename=pv.asStringValue("download.zip");
+			try {
+				HttpServletResponse response=(HttpServletResponse)input;
+				StreamUtil.openZipDownload(response,filename);
+				resultStream=response.getOutputStream();
+			} catch (IOException e) {
+				throw new CollectionException("cannot open download for ["+filename+"]",e);
+			}
+		} else if (input instanceof String) {
+			String filename=(String)input;
+			if (StringUtils.isEmpty(filename)) {
+				throw new CollectionException("input string cannot be empty but must contain a filename");
+			}
+			try {
+				resultStream =new FileOutputStream(filename);
+			} catch (FileNotFoundException e) {
+				throw new CollectionException("cannot create file ["+filename+"] a specified by input message",e);
+			}
+		}
+		if (resultStream==null) {
+			throw new CollectionException("did not find OutputStream or HttpResponse, and could not find filename");
+		}
+		return new ZipWriter(resultStream,writingElement.isCloseOutputstreamOnExit());
 	}
 
 	public void openEntry(String filename) throws CompressionException {
@@ -83,7 +166,8 @@ public class ZipWriter {
 	}
 
 
-	protected void close() throws CompressionException {
+	@Override
+	public void close() throws CompressionException {
 		closeEntry();
 		try {
 			if (isCloseOnExit()) {
@@ -95,6 +179,73 @@ public class ZipWriter {
 			throw new CompressionException("Cannot close ZipStream",e);
 		}
 	}
+
+	@Override
+	public Message writeItem(Message input, PipeLineSession session, ParameterValueList pvl, IZipWritingElement writingElement) throws CollectionException, TimeoutException {
+		String filename = pvl.get(PARAMETER_FILENAME).asStringValue();
+		String charset = writingElement.getCharset();
+		boolean completeFileHeader = writingElement.isCompleteFileHeader();
+		try {
+			ParameterValue pv = pvl.get(PARAMETER_CONTENTS);
+			Message contents = pv!=null ? pv.asMessage() : null;
+			boolean inputIsSource = Message.isNull(contents);
+			if (inputIsSource) {
+				contents = input;
+			}
+			if (completeFileHeader) {
+				writeEntryWithCompletedHeader(filename, contents, writingElement.isCloseInputstreamOnExit(), charset);
+			} else {
+				writeEntry(filename, contents, writingElement.isCloseInputstreamOnExit(), charset);
+			}
+			return inputIsSource ? Message.nullMessage() : input;
+		} catch (IOException | CompressionException e) {
+			throw new CollectionException("cannot write item", e);
+		}
+	}
+
+	@Override
+	public MessageOutputStream provideOutputStream(PipeLineSession session, ParameterValueList pvl, IZipWritingElement writingElement) throws CollectionException {
+		String filename = pvl.get(PARAMETER_FILENAME).asStringValue();
+		String charset = writingElement.getCharset();
+		boolean completeFileHeader = writingElement.isCompleteFileHeader();
+
+		if (completeFileHeader) {
+			return null;
+		}
+		ParameterValue pv = pvl.get(PARAMETER_CONTENTS);
+		Message contents = pv!=null ? pv.asMessage() : null;
+		boolean inputIsSource = Message.isNull(contents);
+		if (!inputIsSource) {
+			return null;
+		}
+		try {
+			openEntry(filename);
+			OutputStream stream = StreamUtil.dontClose(getZipoutput());
+			return new MessageOutputStream(writingElement, stream, (IForwardTarget)null, charset) {
+
+				@Override
+				public void afterClose() throws Exception {
+					stream.flush();
+					closeEntry();
+				}
+
+			};
+		} catch (CompressionException e) {
+			throw new CollectionException("cannot prepare collection to stream item", e);
+		}
+	}
+
+	@Override
+	public OutputStream streamItem(Message input, PipeLineSession session, ParameterValueList pvl, IZipWritingElement writingElement) throws CollectionException {
+		String filename = pvl.get(PARAMETER_FILENAME).asStringValue();
+		try {
+			openEntry(filename);
+		} catch (CompressionException e) {
+			throw new CollectionException("cannot prepare collection to stream item", e);
+		}
+		return StreamUtil.dontClose(getZipoutput());
+	}
+
 
 	public void writeEntry(String filename, Message contents, boolean close, String charset) throws CompressionException, IOException {
 		if (StringUtils.isEmpty(filename)) {
@@ -142,19 +293,5 @@ public class ZipWriter {
 		getZipoutput().closeEntry();
 	}
 
-	public String getLogPrefix(String handlekey) {
-		return "ZipWriterHandle ["+handlekey+"] ";
-	}
-
-	public void setCloseOnExit(boolean b) {
-		closeOnExit = b;
-	}
-	public boolean isCloseOnExit() {
-		return closeOnExit;
-	}
-
-	public ZipOutputStream getZipoutput() {
-		return zipoutput;
-	}
 
 }